requirements.txt
py==1.4.31
pytest==2.9.1
PyYAML==3.11
requests==2.10.0
six==1.10.0
vcrpy==1.7.4
wrapt==1.10.8
